# backend/docs/swagger.yaml
definitions:
  models.Ingredient:
    properties:
      ingredient_description:
        type: string
      ingredient_id:
        type: integer
      ingredient_name:
        type: string
    type: object
  models.IngredientRequest:
    properties:
      ingredient_description:
        type: string
      ingredient_name:
        type: string
    required:
    - ingredient_name
    type: object
...
info:
  contact: {}
paths:
  /ingredients:
    get:
      consumes:
      - application/json
      description: Get all ingredients from the database
      produces:
      - application/json
      responses:
        "200":
          description: List of ingredients
          schema:
            items:
              $ref: '#/definitions/models.Ingredient'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all ingredients
      tags:
      - ingredients
    post:
      consumes:
      - application/json
      description: Add a new ingredient to the database
      parameters:
      - description: Add ingredient
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/models.IngredientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Ingredient'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new ingredient
      tags:
      - ingredients
  /ingredients/{id}:
    delete:
      consumes:
      - application/json
      description: Remove an ingredient by ID from the database
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Ingredient deleted
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete an ingredient
      tags:
      - ingredients
    get:
      consumes:
      - application/json
      description: Get details of a specific ingredient by ID
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ingredient retrieved
          schema:
            $ref: '#/definitions/models.Ingredient'
        "400":
          description: Invalid ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Ingredient not found
          schema:
            additionalProperties: true
            type: object
      summary: Get a single ingredient
      tags:
      - ingredients
    put:
      consumes:
      - application/json
      description: Update an ingredient by ID
      parameters:
      - description: Ingredient ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ingredient content
        in: body
        name: ingredient
        required: true
        schema:
          $ref: '#/definitions/models.Ingredient'
      produces:
      - application/json
      responses:
        "204":
          description: Ingredient updated
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update an existing ingredient
      tags:
      - ingredients
...
swagger: "2.0"